pyproject.toml
[tool.poetry]
name = "mitmproxy"
version = "0.0.1"
description = ""
authors = ["João V. Sanches <joaovictorsanchesandradeestuda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
